Reordered tree species, check desc.
Trees are ordered first by the line they belong to, and then by the tier they belong in. The tier is the number of 'steps' required to get back to a vanilla sapling. (E.G. oak is T0, Lime is T1, Cherry is T2, etc.)

Lines are sorted by:
Temperate
Swamp
Floral
Coniferous
Jungle
Arid